requirements.txt
azure-ai-formrecognizer==3.2.0b3
azure-common==1.1.28
azure-core==1.24.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msrest==0.7.0
oauthlib==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
